Update struct tests in native
