+++ modules/ballerina-test/src/test/resources/test-src/structs/constants/struct-constants.bal
@@ -2,8 +2,8 @@ package constants;
 
 const Person p = 10;
 
-function main(string[] i) {
-	return;
+function main (string[] i) {
+    return;
 }
 
 struct Person {
