+++ modules/ballerina-test/src/test/resources/test-src/structs/duplicate-structs.bal
@@ -1,10 +1,10 @@
 struct Department {
-	string dptName;
-	boolean y;
-	Person prsn1;
+    string dptName;
+    boolean y;
+    Person prsn1;
 }
 
 struct Department {
-	string name;
-	int[] a1;
+    string name;
+    int[] a1;
 }
