+++ modules/ballerina-test/src/test/resources/test-src/structs/invalid-type-attribute-init.bal
@@ -1,9 +1,9 @@
-function testInvalidTypeAttributeinit() {
-	string name;
-	Department dpt = {dptNames:54};
+function testInvalidTypeAttributeinit () {
+    string name;
+    Department dpt = {dptNames:54};
 }
 
 struct Department {
-	map dptNames;
-	int count;
+    map dptNames;
+    int count;
 }
