+++ modules/ballerina-test/src/test/resources/test-src/structs/struct-access-with-dynamic-index.bal
@@ -9,8 +9,8 @@ struct StructField {
     string key;
 }
 
-function testExpressionAsStructIndex() (string){
+function testExpressionAsStructIndex () (string) {
     StructField nameField = {key:"name"};
-    Person emp = {name:"Jack", adrs:{"country":"USA","state":"CA"}, age:25};
+    Person emp = {name:"Jack", adrs:{"country":"USA", "state":"CA"}, age:25};
     return emp[nameField.key];
 }
