+++ modules/ballerina-test/src/test/resources/test-src/structs/undeclared-attribute-access.bal
@@ -1,11 +1,11 @@
-function testUndeclaredAttributeAccess() {
-	string name;
-	Department dpt;
-	
-	dpt.id = "HR";
+function testUndeclaredAttributeAccess () {
+    string name;
+    Department dpt;
+
+    dpt.id = "HR";
 }
 
 struct Department {
-	string dptName;
-	int count;
+    string dptName;
+    int count;
 }
